internationalization to admin user and settings user pages
