+++ src/main/java/com/epam/lowcost/DAO/implementations/FlightDAOImpl.java
@@ -1,8 +1,8 @@
-package com.epam.lowcost.DAO;
+package com.epam.lowcost.DAO.implementations;
 
+import com.epam.lowcost.DAO.interfaces.FlightDAO;
 import com.epam.lowcost.model.Flight;
 import com.epam.lowcost.util.DateFormatter;
-import com.epam.lowcost.util.FlightRowMapper;
 import org.springframework.jdbc.core.RowMapper;
 
 import javax.sql.DataSource;
