+++ src/main/java/com/epam/lowcost/DAO/implementations/UserDAOImpl.java
@@ -1,5 +1,6 @@
-package com.epam.lowcost.DAO;
+package com.epam.lowcost.DAO.implementations;
 
+import com.epam.lowcost.DAO.interfaces.UserDAO;
 import com.epam.lowcost.model.User;
 import com.epam.lowcost.util.DateFormatter;
 import org.springframework.dao.EmptyResultDataAccessException;
