+++ src/main/java/com/epam/lowcost/DAO/interfaces/TicketDAO.java
@@ -1,11 +1,13 @@
-package com.epam.lowcost.service;
+package com.epam.lowcost.DAO.interfaces;
 
 import com.epam.lowcost.model.Ticket;
+
 import java.util.List;
 
-public interface TicketService {
-    List<Ticket> getAllUserTickets(long userId);
+public interface TicketDAO {
     List<Ticket> getAllTickets();
+    List<Ticket> getAllUserTickets(long userId);
+    Ticket getById(long currentId);
     Ticket addTicket(Ticket ticket);
     Ticket updateTicket(Ticket ticket);
     Ticket deleteTicket(long id);
