+++ src/main/java/com/epam/lowcost/controller/LoginController.java
@@ -1,7 +1,7 @@
 package com.epam.lowcost.controller;
 
 import com.epam.lowcost.model.User;
-import com.epam.lowcost.service.UserService;
+import com.epam.lowcost.service.interfaces.UserService;
 import org.mindrot.jbcrypt.BCrypt;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -10,6 +10,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.bind.support.SessionStatus;
 
 import java.util.Map;
+import java.util.ResourceBundle;
 
 
 @Controller
@@ -38,10 +39,12 @@ public class LoginController {
     @PostMapping
     public String login(@RequestParam Map<String, String> logPass, Model model) {
 
+        ResourceBundle bundle = ResourceBundle.getBundle("lang");
+
         User user = userService.verifyUser(logPass.get("email"), logPass.get("password"));
 
         if (user == null) {
-            model.addAttribute("message", "No such User found, or password is wrong. Maybe you want to: ");
+            model.addAttribute("message", bundle.getString("lang.W1NoSuchUser"));
         } else {
             model.addAttribute("sessionUser", user);
             return "redirect:/tickets/self";
@@ -54,5 +57,6 @@ public class LoginController {
         sessionStatus.setComplete();
         return "redirect:/entry";
     }
+
 
 }
