+++ src/main/java/com/epam/lowcost/controller/PlaneController.java
@@ -1,7 +1,7 @@
 package com.epam.lowcost.controller;
 
 import com.epam.lowcost.model.Plane;
-import com.epam.lowcost.service.PlaneService;
+import com.epam.lowcost.service.interfaces.PlaneService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -11,6 +11,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import java.util.Map;
+import java.util.ResourceBundle;
 
 @Controller
 @RequestMapping(value = "/plane")
