+++ src/main/java/com/epam/lowcost/controller/TicketController.java
@@ -3,7 +3,7 @@ package com.epam.lowcost.controller;
 import com.epam.lowcost.model.Flight;
 import com.epam.lowcost.model.Ticket;
 import com.epam.lowcost.model.User;
-import com.epam.lowcost.service.TicketService;
+import com.epam.lowcost.service.interfaces.TicketService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
