+++ src/main/resources/lang_en.properties
@@ -1,0 +1,28 @@
+lang=en
+lang.login=Login
+lang.password=Password
+lang.logIn=Log in
+lang.loginPageEntry=Welcome!
+lang.signIn=Log up?
+lang.logOut=Log out
+lang.currentUser=You - 
+lang.adminPage=Admin Page Current admin is
+lang.W1NoSuchUser=No such User found, or password is wrong. Maybe you want to:
+lang.flightDAO=Edit flight data
+lang.ticketDAO=Edit ticket data
+lang.planeDAO=Edit plane data
+lang.userDAO=Edit user data
+lang.title=AirLin
+lang.loginIntroduction=Enter login and password!
+lang.userSettings=Personal Settings
+lang.backToCabinet=Back to my personal cabinet.
+lang.newEmail=Change Email.
+lang.newFirstName=Change first name.
+lang.newLastName=Change last name.
+lang.newDocument=Change document data.
+lang.newDateBirth=Change birthday.
+lang.btnUpdate=Update
+lang.oldPassword=Old password
+lang.newPassword=New password.
+lang.repeatPassword=Repeat password
+lang.changePersonalData=Change personal data
