+++ src/main/resources/lang_ru.properties
@@ -1,0 +1,28 @@
+lang=ru
+lang.login=Логин
+lang.password=Пароль
+lang.logIn=Вход
+lang.loginPageEntry=Добро пожаловать!
+lang.signIn=Зарегистрироваться?
+lang.logOut=Выйти
+lang.currentUser=Вы -
+lang.adminPage=Страница администратора. Текущий администратор: 
+lang.W1NoSuchUser=Неверное имя пользователя или пароль. Может вы хотите:
+lang.flightDAO=Редактирование данных о полете
+lang.ticketDAO=Редактирование данных о билете
+lang.planeDAO=Редактирование данных о самолетах
+lang.userDAO=Редактирование данных о пользователе
+lang.title=AirLin
+lang.loginIntroduction=Введите логин и пароль!
+lang.userSettings=Редактирование личной информации
+lang.backToCabinet=Назад в личный кабинет.
+lang.newEmail=Изменить Email.
+lang.newFirstName=Изменить имя.
+lang.newLastName=Изменить фамилию.
+lang.newDocument=Изменить паспортные данные.
+lang.newDateBirth=Изменить дату рождения.
+lang.btnUpdate=Обновить
+lang.oldPassword=Старый пароль
+lang.newPassword=Новый пароль.
+lang.repeatPassword=Повторите пароль
+lang.changePersonalData=Редактирование личных данных
