+++ src/main/webapp/WEB-INF/views/flights.jsp
@@ -1,10 +1,12 @@
 <%@taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c" %>
+<%@ taglib prefix="spring" uri="http://www.springframework.org/tags" %>
 <%@ page contentType="text/html;charset=UTF-8" language="java" %>
 <html>
 <head>
-    <title>Flight DAO page Current admin is ${sessionUser.firstName}</title>
+    <title><spring:message code="lang.flightDAO"/></title>
 </head>
 <body>
+<h2><spring:message code="lang.adminPage"/> ${sessionUser.firstName}</h2>
 <h2> Flight in DB:</h2><br/><br/>
 
 <%--<a href="${pageContext.request.contextPath}/flights/all">Show all Flights!</a><br/><br/>--%>
