+++ src/main/webapp/WEB-INF/views/userSettings.jsp
@@ -1,3 +1,4 @@
+<%@ taglib prefix="spring" uri="http://www.springframework.org/tags" %>
 <%--
   Created by IntelliJ IDEA.
   User: Ilia_Stepanov
@@ -8,36 +9,36 @@
 <%@ page contentType="text/html;charset=UTF-8" language="java" %>
 <html>
 <head>
-    <title>User Settings</title>
+    <title><spring:message code="lang.userSettings"/></title>
     <link href="webjars/bootstrap/4.3.1/css/bootstrap-grid.min.css" rel="stylesheet">
-    <p align="right">Current USER in Session: ${sessionUser.firstName}  <a href="/entry/log-out"> Log Out.</a><br/></p>
+    <p align="right"><spring:message code="lang.currentUser"/>  ${sessionUser.firstName}  <a href="/entry/log-out"> <spring:message code="lang.logOut"/></a><br/></p>
 </head>
 <body>
 
-<a href="/self">Back to my page.</a><br/>
+<a href="/tickets/self"><spring:message code="lang.backToCabinet"/></a><br/>
 
-Change personal information.<br/>
+<spring:message code="lang.changePersonalData"/><br/>
 
 <form action="/user/update" method="post">
     <input type="hidden" name="id" value="${sessionUser.id}"/>
-    <input type="email" placeholder="${sessionUser.email}" name="email" /> New Email. <br/>
+    <input type="email" placeholder="${sessionUser.email}" name="email" /> <spring:message code="lang.newEmail"/> <br/>
     <input type="hidden" name="password" value="${sessionUser.password}"/>
     <input type="hidden" name="isAdmin" value="false"/>
-    <input type="text" name="firstName" placeholder="${sessionUser.firstName}"/> New First name. <br/>
-    <input type="text" name="lastName" placeholder="${sessionUser.lastName}"/> New Last name. <br/>
-    <input type="text" name="documentInfo" placeholder="${sessionUser.documentInfo}"/> New Document. <br/>
-    <input type="datetime-local" name="birthday" placeholder="${sessionUser.birthday}"/> New Birthday. <br/>
+    <input type="text" name="firstName" placeholder="${sessionUser.firstName}"/> <spring:message code="lang.newFirstName"/> <br/>
+    <input type="text" name="lastName" placeholder="${sessionUser.lastName}"/> <spring:message code="lang.newLastName"/> <br/>
+    <input type="text" name="documentInfo" placeholder="${sessionUser.documentInfo}"/> <spring:message code="lang.newDocument"/> <br/>
+    <input type="datetime-local" name="birthday" placeholder="${sessionUser.birthday}"/> <spring:message code="lang.newDateBirth"/><br/>
     <input type="hidden" name="userUpdate" value="fromUser"/>
-    <input type="submit" value="UpdateUser"/>
+    <input type="submit" value="ОК"/>
 </form>
 
 
 Change password.
 <form action="/user/change-password" method="post">
-    <input type="password" name="oldPassword"> Old password <br/>
-    <input type="password" name="newPassword"> New password <br/>
-    <input type="password" name="newPassword2"> Repeat new password.<br/>
-    <input type="submit" value="Submit">
+    <input type="password" name="oldPassword"> <spring:message code="lang.oldPassword"/> <br/>
+    <input type="password" name="newPassword"> <spring:message code="lang.newPassword"/><br/>
+    <input type="password" name="newPassword2"> <spring:message code="lang.repeatPassword"/><br/>
+    <input type="submit" value="OK">
 </form>
