+++ src/main/webapp/index.jsp
@@ -3,19 +3,9 @@
 
 <html>
 <head>
-    <title>Title of web app!</title>
+    <title>AirLin</title>
 </head>
 <body>
-
-<h1>Login page.</h1>
-
-<c:redirect url="/entry"/>
-
-<form action="/entry" method="post">
-    <input type="text" name="email" placeholder="Login"/>
-    <input type="password" name="password" placeholder="password"/>
-    <input type="submit" name="Log In!"/>
-</form>
-
+    <c:redirect url="/entry"/>
 </body>
 </html>
